Move documentation under utils
