+++ build.gradle.kts
@@ -80,7 +80,8 @@ configure<com.epages.restdocs.apispec.gradle.OpenApi3Extension> {
     description = apiSpecDescription
     version = "${project.version}"
     format = "json"
-    tagDescriptionsPropertiesFile = "src/test/kotlin/pt/up/fe/ni/website/backend/documentation/tag-descriptions.yaml"
+    tagDescriptionsPropertiesFile =
+        "src/test/kotlin/pt/up/fe/ni/website/backend/utils/documentation/tag-descriptions.yaml"
 }
 
 configure<com.epages.restdocs.apispec.gradle.PostmanExtension> {
