--- src/test/kotlin/pt/up/fe/ni/website/backend/documentation/payloadschemas/EmptyObjectSchema.kt
@@ -1,8 +1,0 @@
-package pt.up.fe.ni.website.backend.documentation.payloadschemas
-
-import pt.up.fe.ni.website.backend.utils.documentation.PayloadSchema
-
-class EmptyObjectSchema : PayloadSchema(
-    "empty",
-    mutableListOf()
-)
