+++ src/test/kotlin/pt/up/fe/ni/website/backend/utils/documentation/Tag.kt
@@ -1,6 +1,6 @@
-package pt.up.fe.ni.website.backend.documentation
+package pt.up.fe.ni.website.backend.utils.documentation
 
-import pt.up.fe.ni.website.backend.utils.documentation.ITag
+import pt.up.fe.ni.website.backend.utils.documentation.utils.ITag
 
 enum class Tag(override val fullName: String) : ITag {
     AUTH("Authentication"),
