+++ src/test/kotlin/pt/up/fe/ni/website/backend/utils/documentation/payloadschemas/EmptyObjectSchema.kt
@@ -1,0 +1,8 @@
+package pt.up.fe.ni.website.backend.utils.documentation.payloadschemas
+
+import pt.up.fe.ni.website.backend.utils.documentation.utils.PayloadSchema
+
+class EmptyObjectSchema : PayloadSchema(
+    "empty",
+    mutableListOf()
+)
