+++ src/test/kotlin/pt/up/fe/ni/website/backend/utils/documentation/payloadschemas/ErrorSchema.kt
@@ -1,8 +1,8 @@
-package pt.up.fe.ni.website.backend.documentation.payloadschemas
+package pt.up.fe.ni.website.backend.utils.documentation.payloadschemas
 
 import org.springframework.restdocs.payload.JsonFieldType
-import pt.up.fe.ni.website.backend.utils.documentation.DocumentedJSONField
-import pt.up.fe.ni.website.backend.utils.documentation.PayloadSchema
+import pt.up.fe.ni.website.backend.utils.documentation.utils.DocumentedJSONField
+import pt.up.fe.ni.website.backend.utils.documentation.utils.PayloadSchema
 
 class ErrorSchema : PayloadSchema(
     "error",
