+++ src/test/kotlin/pt/up/fe/ni/website/backend/utils/documentation/payloadschemas/model/PayloadProject.kt
@@ -1,10 +1,10 @@
-package pt.up.fe.ni.website.backend.documentation.payloadschemas.model
+package pt.up.fe.ni.website.backend.utils.documentation.payloadschemas.model
 
 import org.springframework.restdocs.payload.JsonFieldType
-import pt.up.fe.ni.website.backend.documentation.Tag
-import pt.up.fe.ni.website.backend.utils.documentation.DocumentedJSONField
-import pt.up.fe.ni.website.backend.utils.documentation.DocumentedJSONField.Companion.addFieldsBeneathPath
-import pt.up.fe.ni.website.backend.utils.documentation.ModelDocumentation
+import pt.up.fe.ni.website.backend.utils.documentation.Tag
+import pt.up.fe.ni.website.backend.utils.documentation.utils.DocumentedJSONField
+import pt.up.fe.ni.website.backend.utils.documentation.utils.DocumentedJSONField.Companion.addFieldsBeneathPath
+import pt.up.fe.ni.website.backend.utils.documentation.utils.ModelDocumentation
 
 class PayloadProject : ModelDocumentation(
     Tag.PROJECT.name.lowercase(),
